+++ src/register.js
@@ -1,14 +1,9 @@
 const LocalStorage = require("./LocalStorage");
 
-const COMMANDS_PREFIX = "localStorage";
-
 const register = (Cypress, localStorage) => {
   const localStorageCommands = new LocalStorage(localStorage);
-  localStorageCommands.publicMethods.forEach(methodName => {
-    Cypress.Commands.add(
-      `${COMMANDS_PREFIX}${methodName}`,
-      localStorageCommands[methodName].bind(localStorageCommands)
-    );
+  LocalStorage.publicMethods.forEach(methodName => {
+    Cypress.Commands.add(methodName, localStorageCommands[methodName].bind(localStorageCommands));
   });
 };